convert DatabaseList to typescript
