+++ frontend/src/metabase/admin/databases/components/DatabaseList/DatabaseList.tsx
@@ -1,0 +1,145 @@
+import cx from "classnames";
+import { Link } from "react-router";
+import { t } from "ttag";
+
+import LoadingSpinner from "metabase/components/LoadingSpinner";
+import AdminS from "metabase/css/admin.module.css";
+import ButtonsS from "metabase/css/components/buttons.module.css";
+import CS from "metabase/css/core/index.css";
+import { FormMessage } from "metabase/forms";
+import { isSyncCompleted } from "metabase/lib/syncing";
+import { PLUGIN_FEATURE_LEVEL_PERMISSIONS } from "metabase/plugins";
+import type { Database, Engine } from "metabase-types/api";
+
+import {
+  AddSampleDatabaseLink,
+  TableCellContent,
+  TableCellSpinner,
+} from "../../containers/DatabaseListApp.styled";
+
+const query = {
+  ...PLUGIN_FEATURE_LEVEL_PERMISSIONS.databaseDetailsQueryProps,
+};
+
+interface DatabaseListProps {
+  databases: Database[];
+  hasSampleDatabase: boolean;
+  engines: Record<string, Engine>;
+  deletes: Array<any>;
+  deletionError: any;
+  isAdmin: boolean;
+  isAddingSampleDatabase: boolean;
+  addSampleDatabase: (query: any) => void;
+  addSampleDatabaseError: boolean;
+  children: React.ReactNode;
+}
+
+export const DatabaseList = ({
+  children,
+  databases,
+  addSampleDatabase,
+  hasSampleDatabase,
+  isAddingSampleDatabase,
+  addSampleDatabaseError,
+  engines,
+  deletes,
+  deletionError,
+  isAdmin,
+}: DatabaseListProps) => {
+  const error = deletionError || addSampleDatabaseError;
+
+  return (
+    <>
+      <div className={CS.wrapper} data-testid="database-list">
+        <section className={cx(AdminS.PageHeader, CS.px2, CS.clearfix)}>
+          {isAdmin && (
+            <Link
+              to="/admin/databases/create"
+              className={cx(
+                ButtonsS.Button,
+                ButtonsS.ButtonPrimary,
+                CS.floatRight,
+              )}
+            >{t`Add database`}</Link>
+          )}
+          <h2 className={CS.m0}>{t`Databases`}</h2>
+        </section>
+        {error && (
+          <section>
+            <FormMessage formError={error} />
+          </section>
+        )}
+        <section>
+          <table className={AdminS.ContentTable}>
+            <thead>
+              <tr>
+                <th>{t`Name`}</th>
+                <th>{t`Engine`}</th>
+              </tr>
+            </thead>
+            <tbody>
+              {databases ? (
+                databases.map((database) => (
+                  <tr
+                    key={database.id}
+                    className={cx({
+                      disabled: deletes.indexOf(database.id) !== -1,
+                    })}
+                  >
+                    <td>
+                      <TableCellContent>
+                        {!isSyncCompleted(database) && (
+                          <TableCellSpinner size={16} borderWidth={2} />
+                        )}
+                        <Link
+                          to={"/admin/databases/" + database.id}
+                          className={cx(CS.textBold, CS.link)}
+                        >
+                          {database.name}
+                        </Link>
+                      </TableCellContent>
+                    </td>
+                    <td>
+                      {engines?.[database.engine ?? ""]?.["driver-name"] ??
+                        database.engine}
+                    </td>
+                  </tr>
+                ))
+              ) : (
+                <tr>
+                  <td colSpan={4}>
+                    <LoadingSpinner />
+                    <h3>{t`Loading ...`}</h3>
+                  </td>
+                </tr>
+              )}
+            </tbody>
+          </table>
+          {!hasSampleDatabase && isAdmin ? (
+            <div className={CS.pt4}>
+              <span
+                className={cx(CS.p2, CS.textItalic, {
+                  [CS.borderTop]: databases && databases.length > 0,
+                })}
+              >
+                {isAddingSampleDatabase ? (
+                  <span className={cx(CS.textLight, CS.noDecoration)}>
+                    {t`Restoring the sample database...`}
+                  </span>
+                ) : (
+                  <AddSampleDatabaseLink
+                    onClick={() => addSampleDatabase(query)}
+                  >
+                    {t`Bring the sample database back`}
+                  </AddSampleDatabaseLink>
+                )}
+              </span>
+            </div>
+          ) : null}
+        </section>
+      </div>
+
+      {children}
+    </>
+  );
+};
