+++ frontend/src/metabase/admin/databases/components/DatabaseList/index.ts
@@ -1,2 +1,1 @@
-// eslint-disable-next-line import/no-default-export -- deprecated usage
-export { default } from "./DatabaseList";
+export * from "./DatabaseList";
